+++ Source/DlgSystem/Private/DlgContext.cpp
@@ -4,6 +4,7 @@
 #include "DlgSystemPrivatePCH.h"
 #include "Nodes/DlgNode.h"
 #include "Nodes/DlgNode_End.h"
+#include "Nodes/DlgNode_SpeechSequence.h"
 #include "DlgDialogueParticipant.h"
 #include "DlgMemory.h"
 #include "Engine/Texture2D.h"
@@ -45,7 +46,7 @@ void UDlgContext::OnRep_SerializedParticipants()
 	}
 }
 
-bool UDlgContext::ChooseChild(int32 OptionIndex)
+bool UDlgContext::ChooseOption(int32 OptionIndex)
 {
 	check(Dialogue);
 	if (UDlgNode* Node = GetMutableActiveNode())
@@ -60,18 +61,33 @@ bool UDlgContext::ChooseChild(int32 OptionIndex)
 	return false;
 }
 
-bool UDlgContext::ChooseChildBasedOnAllOptionIndex(int32 Index)
+bool UDlgContext::ChooseSpeechSequenceOptionFromReplicated(int32 OptionIndex)
+{
+	check(Dialogue);
+	if (UDlgNode_SpeechSequence* Node = GetMutableActiveNodeAsSpeechSequence())
+	{
+		if (Node->OptionSelectedFromReplicated(OptionIndex, *this))
+		{
+			return true;
+		}
+	}
+
+	bDialogueEnded = true;
+	return false;
+}
+
+bool UDlgContext::ChooseOptionBasedOnAllOptionIndex(int32 Index)
 {
 	if (!AllChildren.IsValidIndex(Index))
 	{
-		LogErrorWithContext(FString::Printf(TEXT("ChooseChildBasedOnAllOptionIndex - INVALID given Index = %d"), Index));
+		LogErrorWithContext(FString::Printf(TEXT("ChooseOptionBasedOnAllOptionIndex - INVALID given Index = %d"), Index));
 		bDialogueEnded = true;
 		return false;
 	}
 
 	if (!AllChildren[Index].IsSatisfied())
 	{
-		LogErrorWithContext(FString::Printf(TEXT("ChooseChildBasedOnAllOptionIndex - given Index = %d is an unsatisfied edge"), Index));
+		LogErrorWithContext(FString::Printf(TEXT("ChooseOptionBasedOnAllOptionIndex - given Index = %d is an unsatisfied edge"), Index));
 		bDialogueEnded = true;
 		return false;
 	}
@@ -80,7 +96,7 @@ bool UDlgContext::ChooseChildBasedOnAllOptionIndex(int32 Index)
 	{
 		if (AvailableChildren[i] == AllChildren[Index].GetEdge())
 		{
-			return ChooseChild(i);
+			return ChooseOption(i);
 		}
 	}
 
@@ -89,17 +105,17 @@ bool UDlgContext::ChooseChildBasedOnAllOptionIndex(int32 Index)
 	return false;
 }
 
-void UDlgContext::ReevaluateChildren()
+bool UDlgContext::ReevaluateOptions()
 {
 	check(Dialogue);
 	UDlgNode* Node = GetMutableActiveNode();
 	if (!IsValid(Node))
 	{
-		LogErrorWithContext(TEXT("ReevaluateChildren - Failed to update dialogue options"));
-		return;
+		LogErrorWithContext(TEXT("ReevaluateOptions - Failed to update dialogue options"));
+		return false;
 	}
 
-	Node->ReevaluateChildren(*this, {});
+	return Node->ReevaluateChildren(*this, {});
 }
 
 const FText& UDlgContext::GetOptionText(int32 OptionIndex) const
@@ -320,9 +336,9 @@ UObject* UDlgContext::GetActiveNodeParticipant() const
 	if (ObjectPtr == nullptr || !IsValid(*ObjectPtr))
 	{
 		LogErrorWithContext(FString::Printf(
-            TEXT("GetActiveNodeParticipant - The ParticipantName = `%s` from the Active Node does NOT exist in the current Participants"),
-            *SpeakerName.ToString()
-        ));
+			TEXT("GetActiveNodeParticipant - The ParticipantName = `%s` from the Active Node does NOT exist in the current Participants"),
+			*SpeakerName.ToString()
+		));
 		return nullptr;
 	}
 
@@ -363,6 +379,26 @@ const UObject* UDlgContext::GetParticipant(FName ParticipantName) const
 	return nullptr;
 }
 
+bool UDlgContext::IsValidNodeIndex(int32 NodeIndex) const
+{
+	return Dialogue ? Dialogue->IsValidNodeIndex(NodeIndex) : false;
+}
+
+bool UDlgContext::IsValidNodeGUID(const FGuid& NodeGUID) const
+{
+	return Dialogue ? Dialogue->IsValidNodeGUID(NodeGUID) : false;
+}
+
+FGuid UDlgContext::GetNodeGUIDForIndex(int32 NodeIndex) const
+{
+	return Dialogue ? Dialogue->GetNodeGUIDForIndex(NodeIndex) : FGuid{};
+}
+
+int32 UDlgContext::GetNodeIndexForGUID(const FGuid& NodeGUID) const
+{
+	return Dialogue ? Dialogue->GetNodeIndexForGUID(NodeGUID) : INDEX_NONE;
+}
+
 bool UDlgContext::IsOptionConnectedToVisitedNode(int32 Index, bool bLocalHistory, bool bIndexSkipsUnsatisfiedEdges) const
 {
 	int32 TargetIndex = INDEX_NONE;
@@ -386,9 +422,10 @@ bool UDlgContext::IsOptionConnectedToVisitedNode(int32 Index, bool bLocalHistory
 		TargetIndex = AllChildren[Index].GetEdge().TargetIndex;
 	}
 
+	const FGuid TargetGUID = GetNodeGUIDForIndex(TargetIndex);
 	if (bLocalHistory)
 	{
-		return VisitedNodeIndices.Contains(TargetIndex);
+		return History.Contains(TargetIndex, TargetGUID);
 	}
 
 	if (Dialogue == nullptr)
@@ -397,7 +434,7 @@ bool UDlgContext::IsOptionConnectedToVisitedNode(int32 Index, bool bLocalHistory
 		return false;
 	}
 
-	return FDlgMemory::Get().IsNodeVisited(Dialogue->GetDialogueGUID(), TargetIndex);
+	return FDlgMemory::Get().IsNodeVisited(Dialogue->GetGUID(), TargetIndex, TargetGUID);
 }
 
 bool UDlgContext::IsOptionConnectedToEndNode(int32 Index, bool bIndexSkipsUnsatisfiedEdges) const
@@ -442,7 +479,7 @@ bool UDlgContext::IsOptionConnectedToEndNode(int32 Index, bool bIndexSkipsUnsati
 bool UDlgContext::EnterNode(int32 NodeIndex, TSet<const UDlgNode*> NodesEnteredWithThisStep)
 {
 	check(Dialogue);
-	UDlgNode* Node = GetMutableNode(NodeIndex);
+	UDlgNode* Node = GetMutableNodeFromIndex(NodeIndex);
 	if (!IsValid(Node))
 	{
 		LogErrorWithContext(FString::Printf(TEXT("EnterNode - FAILED because of INVALID NodeIndex = %d"), NodeIndex));
@@ -450,40 +487,75 @@ bool UDlgContext::EnterNode(int32 NodeIndex, TSet<const UDlgNode*> NodesEnteredW
 	}
 
 	ActiveNodeIndex = NodeIndex;
-	FDlgMemory::Get().SetNodeVisited(Dialogue->GetDialogueGUID(), ActiveNodeIndex);
-	VisitedNodeIndices.Add(ActiveNodeIndex);
+	SetNodeVisited(NodeIndex, Node->GetGUID());
 
 	return Node->HandleNodeEnter(*this, NodesEnteredWithThisStep);
 }
 
-UDlgNode* UDlgContext::GetMutableNode(int32 NodeIndex) const
+void UDlgContext::SetNodeVisited(int32 NodeIndex, const FGuid& NodeGUID)
 {
-	check(Dialogue);
-	const TArray<UDlgNode*>& Nodes = Dialogue->GetNodes();
-	if (!Nodes.IsValidIndex(NodeIndex))
-	{
-		return nullptr;
-	}
-
-	return Nodes[NodeIndex];
+	FDlgMemory::Get().SetNodeVisited(Dialogue->GetGUID(), NodeIndex, NodeGUID);
+	History.Add(NodeIndex, NodeGUID);
 }
 
-const UDlgNode* UDlgContext::GetNode(int32 NodeIndex) const
+UDlgNode_SpeechSequence* UDlgContext::GetMutableActiveNodeAsSpeechSequence() const
+{
+	return Cast<UDlgNode_SpeechSequence>(GetMutableNodeFromIndex(ActiveNodeIndex));
+}
+
+const UDlgNode_SpeechSequence* UDlgContext::GetActiveNodeAsSpeechSequence() const
+{
+	return Cast<UDlgNode_SpeechSequence>(GetNodeFromIndex(ActiveNodeIndex));
+}
+
+UDlgNode* UDlgContext::GetMutableNodeFromIndex(int32 NodeIndex) const
 {
 	check(Dialogue);
-	const TArray<UDlgNode*>& Nodes = Dialogue->GetNodes();
-	if (!Nodes.IsValidIndex(NodeIndex))
+	if (!Dialogue->IsValidNodeIndex(NodeIndex))
 	{
 		return nullptr;
 	}
 
-	return Nodes[NodeIndex];
+	return Dialogue->GetMutableNodeFromIndex(NodeIndex);
+}
+
+const UDlgNode* UDlgContext::GetNodeFromIndex(int32 NodeIndex) const
+{
+	check(Dialogue);
+	if (!Dialogue->IsValidNodeIndex(NodeIndex))
+	{
+		return nullptr;
+	}
+
+	return Dialogue->GetMutableNodeFromIndex(NodeIndex);
+}
+
+UDlgNode* UDlgContext::GetMutableNodeFromGUID(const FGuid& NodeGUID) const
+{
+	check(Dialogue);
+	if (!Dialogue->IsValidNodeGUID(NodeGUID))
+	{
+		return nullptr;
+	}
+
+	return Dialogue->GetMutableNodeFromGUID(NodeGUID);
+}
+
+const UDlgNode* UDlgContext::GetNodeFromGUID(const FGuid& NodeGUID) const
+{
+	check(Dialogue);
+	if (!Dialogue->IsValidNodeGUID(NodeGUID))
+	{
+		return nullptr;
+	}
+
+	return Dialogue->GetMutableNodeFromGUID(NodeGUID);
 }
 
 bool UDlgContext::IsNodeEnterable(int32 NodeIndex, TSet<const UDlgNode*> AlreadyVisitedNodes) const
 {
 	check(Dialogue);
-	if (const UDlgNode* Node = GetNode(NodeIndex))
+	if (const UDlgNode* Node = GetNodeFromIndex(NodeIndex))
 	{
 		return Node->CheckNodeEnterConditions(*this, AlreadyVisitedNodes);
 	}
@@ -528,7 +600,7 @@ bool UDlgContext::CanBeStarted(UDlgDialogue* InDialogue, const TMap<FName, UObje
 	return false;
 }
 
-bool UDlgContext::StartFromContext(const FString& ContextString, UDlgDialogue* InDialogue, const TMap<FName, UObject*>& InParticipants)
+bool UDlgContext::StartWithContext(const FString& ContextString, UDlgDialogue* InDialogue, const TMap<FName, UObject*>& InParticipants)
 {
 	const FString ContextMessage = ContextString.IsEmpty()
 		? TEXT("Start")
@@ -555,51 +627,57 @@ bool UDlgContext::StartFromContext(const FString& ContextString, UDlgDialogue* I
 	}
 
 	LogErrorWithContext(FString::Printf(
-        TEXT("%s - FAILED because all possible start node condition failed. Edge conditions and children enter conditions from the start node are not satisfied"),
-        *ContextMessage
-    ));
+		TEXT("%s - FAILED because all possible start node condition failed. Edge conditions and children enter conditions from the start node are not satisfied"),
+		*ContextMessage
+	));
 	return false;
 }
 
-bool UDlgContext::StartFromContextFromIndex(
+bool UDlgContext::StartWithContextFromNode(
 	const FString& ContextString,
 	UDlgDialogue* InDialogue,
 	const TMap<FName, UObject*>& InParticipants,
-	int32 StartIndex,
-	const TSet<int32>& VisitedNodes,
+	int32 StartNodeIndex,
+	const FGuid& StartNodeGUID,
+	const FDlgHistory& StartHistory,
 	bool bFireEnterEvents
 )
 {
 	const FString ContextMessage = ContextString.IsEmpty()
-        ? TEXT("StartFromIndex")
-        : FString::Printf(TEXT("%s - StartFromIndex"), *ContextString);
+		? TEXT("StartFromNode")
+		: FString::Printf(TEXT("%s - StartFromNode"), *ContextString);
 
 	Dialogue = InDialogue;
 	SetParticipants(InParticipants);
-	VisitedNodeIndices = VisitedNodes;
+	History = StartHistory;
 	if (!ValidateParticipantsMapForDialogue(ContextMessage, Dialogue, Participants))
 	{
 		return false;
 	}
 
-	UDlgNode* Node = GetMutableNode(StartIndex);
+	// Get the StartNodeIndex from the GUID
+	if (StartNodeGUID.IsValid())
+	{
+		StartNodeIndex = GetNodeIndexForGUID(StartNodeGUID);
+	}
+
+	UDlgNode* Node = GetMutableNodeFromIndex(StartNodeIndex);
 	if (!IsValid(Node))
 	{
 		LogErrorWithContext(FString::Printf(
-			TEXT("%s - FAILED because StartIndex = %d is an INVALID index"),
-			*ContextMessage, StartIndex
+			TEXT("%s - FAILED because StartNodeIndex = %d  is INVALID. For StartNodeGUID = %s"),
+			*ContextMessage, StartNodeIndex, *StartNodeGUID.ToString()
 		));
 		return false;
 	}
 
 	if (bFireEnterEvents)
 	{
-		return EnterNode(StartIndex, {});
+		return EnterNode(StartNodeIndex, {});
 	}
 
-	ActiveNodeIndex = StartIndex;
-	FDlgMemory::Get().SetNodeVisited(Dialogue->GetDialogueGUID(), ActiveNodeIndex);
-	VisitedNodeIndices.Add(ActiveNodeIndex);
+	ActiveNodeIndex = StartNodeIndex;
+	SetNodeVisited(StartNodeIndex, Node->GetGUID());
 
 	return Node->ReevaluateChildren(*this, {});
 }
@@ -614,11 +692,11 @@ FString UDlgContext::GetContextString() const
 	}
 
 	return FString::Printf(
-        TEXT("Dialogue = `%s`, ActiveNodeIndex = %d, Participants Names = `%s`"),
-        Dialogue ? *Dialogue->GetPathName() : TEXT("INVALID"),
-        ActiveNodeIndex,
-        *FString::Join(ParticipantsNames, TEXT(", "))
-    );
+		TEXT("Dialogue = `%s`, ActiveNodeIndex = %d, Participants Names = `%s`"),
+		Dialogue ? *Dialogue->GetPathName() : TEXT("INVALID"),
+		ActiveNodeIndex,
+		*FString::Join(ParticipantsNames, TEXT(", "))
+	);
 }
 
 void UDlgContext::LogErrorWithContext(const FString& ErrorMessage) const
@@ -688,30 +766,30 @@ bool UDlgContext::ValidateParticipantForDialogue(
 
 		case EDlgValidateStatus::DialogueIsNull:
 			FDlgLogger::Get().Errorf(
-			    TEXT("%s - Dialogue is INVALID (not set or null).\nContext:\n\tParticipant = `%s`"),
-			    *ContextString, Participant ? *Participant->GetPathName() : TEXT("INVALID")
+				TEXT("%s - Dialogue is INVALID (not set or null).\nContext:\n\tParticipant = `%s`"),
+				*ContextString, Participant ? *Participant->GetPathName() : TEXT("INVALID")
 			);
 			return false;
 
 		case EDlgValidateStatus::ParticipantIsNull:
 			FDlgLogger::Get().Errorf(
-	            TEXT("%s - Participant is INVALID (not set or null).\nContext:\n\tDialogue = `%s`"),
-	            *ContextString, Dialogue ? *Dialogue->GetPathName() : TEXT("INVALID")
-	        );
+				TEXT("%s - Participant is INVALID (not set or null).\nContext:\n\tDialogue = `%s`"),
+				*ContextString, Dialogue ? *Dialogue->GetPathName() : TEXT("INVALID")
+			);
 			return false;
 
 		case EDlgValidateStatus::ParticipantDoesNotImplementInterface:
 			FDlgLogger::Get().Errorf(
-	            TEXT("%s - Participant Path = `%s` does not implement the IDlgDialogueParticipant/UDlgDialogueParticipant interface.\nContext:\n\tDialogue = `%s`"),
-	            *ContextString, *Participant->GetPathName(), *Dialogue->GetPathName()
-	        );
+				TEXT("%s - Participant Path = `%s` does not implement the IDlgDialogueParticipant/UDlgDialogueParticipant interface.\nContext:\n\tDialogue = `%s`"),
+				*ContextString, *Participant->GetPathName(), *Dialogue->GetPathName()
+			);
 			return false;
 
 		case EDlgValidateStatus::ParticipantIsABlueprintClassAndDoesNotImplementInterface:
-	        FDlgLogger::Get().Errorf(
-	            TEXT("%s - Participant Path = `%s` is a Blueprint Class (from the content browser) and NOT a Blueprint Instance (from the level world).\nContext:\n\tDialogue = `%s`"),
-	            *ContextString, *Participant->GetPathName(), *Dialogue->GetPathName()
-	        );
+			FDlgLogger::Get().Errorf(
+				TEXT("%s - Participant Path = `%s` is a Blueprint Class (from the content browser) and NOT a Blueprint Instance (from the level world).\nContext:\n\tDialogue = `%s`"),
+				*ContextString, *Participant->GetPathName(), *Dialogue->GetPathName()
+			);
 			return false;
 
 		// case EDlgValidateStatus::DialogueDoesNotContainParticipant:
@@ -735,8 +813,8 @@ bool UDlgContext::ValidateParticipantsMapForDialogue(
 )
 {
 	const FString ContextMessage = ContextString.IsEmpty()
-        ? FString::Printf(TEXT("ValidateParticipantsMapForDialogue"))
-        : FString::Printf(TEXT("%s - ValidateParticipantsMapForDialogue"), *ContextString);
+		? FString::Printf(TEXT("ValidateParticipantsMapForDialogue"))
+		: FString::Printf(TEXT("%s - ValidateParticipantsMapForDialogue"), *ContextString);
 
 	if (!IsValid(Dialogue))
 	{
@@ -783,8 +861,8 @@ bool UDlgContext::ValidateParticipantsMapForDialogue(
 				if (bLog)
 				{
 					FDlgLogger::Get().Errorf(
-					    TEXT("%s - The Map has a KEY Participant Name = `%s` DIFFERENT to the VALUE of the Participant Path = `%s` with the Name = `%s` (KEY Participant Name != VALUE Participant Name)"),
-					    *ContextMessage, *ParticipantName.ToString(), *Participant->GetPathName(), *ObjectParticipantName.ToString()
+						TEXT("%s - The Map has a KEY Participant Name = `%s` DIFFERENT to the VALUE of the Participant Path = `%s` with the Name = `%s` (KEY Participant Name != VALUE Participant Name)"),
+						*ContextMessage, *ParticipantName.ToString(), *Participant->GetPathName(), *ObjectParticipantName.ToString()
 					);
 				}
 				return false;
@@ -802,9 +880,9 @@ bool UDlgContext::ValidateParticipantsMapForDialogue(
 			if (bLog)
 			{
 				FDlgLogger::Get().Warningf(
-                    TEXT("%s - Participant Path = `%s` with Participant Name = `%s` is NOT referenced (DOES) not exist inside the Dialogue. It is going to be IGNORED.\nContext:\n\tDialogue = `%s`"),
-                    *ContextMessage, *Participant->GetPathName(), *ParticipantName.ToString(), *Dialogue->GetPathName()
-                );
+					TEXT("%s - Participant Path = `%s` with Participant Name = `%s` is NOT referenced (DOES) not exist inside the Dialogue. It is going to be IGNORED.\nContext:\n\tDialogue = `%s`"),
+					*ContextMessage, *Participant->GetPathName(), *ParticipantName.ToString(), *Dialogue->GetPathName()
+				);
 			}
 		}
 	}
@@ -822,9 +900,9 @@ bool UDlgContext::ValidateParticipantsMapForDialogue(
 
 			const FString NameList = FString::Join(ParticipantsMissing, TEXT(", "));
 			FDlgLogger::Get().Errorf(
-                TEXT("%s - FAILED for Dialogue = `%s` because the following Participant Names are MISSING: `%s"),
-                *ContextMessage,  *Dialogue->GetPathName(), *NameList
-            );
+				TEXT("%s - FAILED for Dialogue = `%s` because the following Participant Names are MISSING: `%s"),
+				*ContextMessage,  *Dialogue->GetPathName(), *NameList
+			);
 		}
 		return false;
 	}
@@ -833,16 +911,16 @@ bool UDlgContext::ValidateParticipantsMapForDialogue(
 }
 
 bool UDlgContext::ConvertArrayOfParticipantsToMap(
-    const FString& ContextString,
-    const UDlgDialogue* Dialogue,
-    const TArray<UObject*>& ParticipantsArray,
-    TMap<FName, UObject*>& OutParticipantsMap,
-    bool bLog
+	const FString& ContextString,
+	const UDlgDialogue* Dialogue,
+	const TArray<UObject*>& ParticipantsArray,
+	TMap<FName, UObject*>& OutParticipantsMap,
+	bool bLog
 )
 {
 	const FString ContextMessage = ContextString.IsEmpty()
-        ? FString::Printf(TEXT("ConvertArrayOfParticipantsToMap"))
-        : FString::Printf(TEXT("%s - ConvertArrayOfParticipantsToMap"), *ContextString);
+		? FString::Printf(TEXT("ConvertArrayOfParticipantsToMap"))
+		: FString::Printf(TEXT("%s - ConvertArrayOfParticipantsToMap"), *ContextString);
 
 	// We don't allow to convert empty arrays
 	OutParticipantsMap.Empty();
@@ -851,8 +929,8 @@ bool UDlgContext::ConvertArrayOfParticipantsToMap(
 		if (bLog)
 		{
 			FDlgLogger::Get().Errorf(
-			    TEXT("%s - Participants Array is EMPTY, can't convert anything. Dialogue = `%s`"),
-			    *ContextMessage, Dialogue ? *Dialogue->GetPathName() : TEXT("INVALID")
+				TEXT("%s - Participants Array is EMPTY, can't convert anything. Dialogue = `%s`"),
+				*ContextMessage, Dialogue ? *Dialogue->GetPathName() : TEXT("INVALID")
 			);
 		}
 		return false;
@@ -877,8 +955,8 @@ bool UDlgContext::ConvertArrayOfParticipantsToMap(
 			if (bLog)
 			{
 				FDlgLogger::Get().Warningf(
-				    TEXT("%s - Participant Path = `%s`, Participant Name = `%s` already exists in the Array. Ignoring it!"),
-				    *ContextMessageWithIndex, *Participant->GetPathName(), *ParticipantName.ToString()
+					TEXT("%s - Participant Path = `%s`, Participant Name = `%s` already exists in the Array. Ignoring it!"),
+					*ContextMessageWithIndex, *Participant->GetPathName(), *ParticipantName.ToString()
 				);
 			}
 			continue;