+++ Source/DlgSystemEditor/Private/DialogueEditor/DetailsPanel/DialogueCondition_Details.cpp
@@ -12,6 +12,7 @@
 #include "IPropertyUtilities.h"
 #include "Widgets/DialogueTextPropertyPickList_CustomRowHelper.h"
 #include "DlgHelper.h"
+#include "Widgets/DialogueIntTextBox_CustomRowHelper.h"
 
 #define LOCTEXT_NAMESPACE "DialogueCondition_Details"
 
@@ -39,7 +40,6 @@ void FDialogueCondition_Details::CustomizeHeader(TSharedRef<IPropertyHandle> InS
 
 	// Register handler properties changes
 	ConditionTypePropertyHandle->SetOnPropertyValueChanged(FSimpleDelegate::CreateSP(this, &Self::OnConditionTypeChanged, true));
-
 	CompareTypePropertyHandle->SetOnPropertyValueChanged(FSimpleDelegate::CreateSP(this, &Self::OnCompareTypeChanged, true));
 
 	const bool bShowOnlyInnerProperties = StructPropertyHandle->GetProperty()->HasMetaData(META_ShowOnlyInnerProperties);
@@ -157,8 +157,11 @@ void FDialogueCondition_Details::CustomizeChildren(TSharedRef<IPropertyHandle> I
 
 	// IntValue
 	{
-		IntValuePropertyRow = &StructBuilder.AddProperty(IntValuePropertyHandle.ToSharedRef());
-		IntValuePropertyRow->Visibility(CREATE_VISIBILITY_CALLBACK(&Self::GetIntValueVisibility));
+		FDetailWidgetRow* DetailWidgetRow = &StructBuilder.AddCustomRow(LOCTEXT("IntValueSearchKey", "Int Value"));
+		IntValuePropertyRow = MakeShared<FDialogueIntTextBox_CustomRowHelper>(DetailWidgetRow, IntValuePropertyHandle, Dialogue);
+		IntValuePropertyRow->SetVisibility(CREATE_VISIBILITY_CALLBACK(&Self::GetIntValueVisibility));
+		IntValuePropertyRow->SetJumpToNodeVisibility(CREATE_VISIBILITY_CALLBACK(&Self::GetGUIDVisibility));
+		IntValuePropertyRow->Update();
 	}
 
 	// FloatValue
@@ -193,6 +196,14 @@ void FDialogueCondition_Details::CustomizeChildren(TSharedRef<IPropertyHandle> I
 		LongTermMemoryPropertyRow->Visibility(CREATE_VISIBILITY_CALLBACK(&Self::GetLongTermMemoryVisibility));
 	}
 
+	// GUID
+	{
+		GUIDPropertyRow = &StructBuilder.AddProperty(
+            StructPropertyHandle->GetChildHandle(GET_MEMBER_NAME_CHECKED(FDlgCondition, GUID)).ToSharedRef()
+        );
+		GUIDPropertyRow->Visibility(CREATE_VISIBILITY_CALLBACK(&Self::GetGUIDVisibility));
+	}
+
 	// CustomCondition
 	{
 		CustomConditionPropertyRow = &StructBuilder.AddProperty(
@@ -222,11 +233,22 @@ void FDialogueCondition_Details::OnConditionTypeChanged(bool bForceRefresh)
 	FText BoolValueDisplayName = LOCTEXT("BoolValueDisplayName", "Return Value");
 	FText BoolValueToolTip = LOCTEXT("BoolValueToolTip", "SHOULD NOT BE VISIBLE");
 	// TODO remove the "equal" operations for float values as they are imprecise
-	FText FloatValueToolTip = LOCTEXT("FloatValueToolTip", "The float value the VariableName is checked against (depending on the operation).\n"
-		"VariableName <Operation> FloatValue");
+	FText FloatValueToolTip = LOCTEXT(
+		"FloatValueToolTip",
+		"The float value the VariableName is checked against (depending on the operation).\n"
+		"VariableName <Operation> FloatValue"
+	);
 	FText IntValueDisplayName = LOCTEXT("IntValueDisplayName", "Int Value");
-	FText IntValueToolTip = LOCTEXT("IntValueToolTip", "The int value the VariableName is checked against (depending on the operation).\n"
-		"VariableName <Operation> IntValue");
+	FText IntValueToolTip = LOCTEXT(
+		"IntValueToolTip",
+		"The int value the VariableName is checked against (depending on the operation).\n"
+		"VariableName <Operation> IntValue"
+	);
+	FText GUIDDisplayName = LOCTEXT("GUIDDisplayName", "Node GUID");
+	FText GUIDToolTip = LOCTEXT(
+		"GUIDToolTip",
+		"The Corresponding GUID of the Node Index. (Set On Compile)"
+	);
 
 	FDialogueDetailsPanelUtils::ResetNumericPropertyLimits(IntValuePropertyHandle);
 	switch (ConditionType)
@@ -294,9 +316,13 @@ void FDialogueCondition_Details::OnConditionTypeChanged(bool bForceRefresh)
 	BoolValuePropertyRow->DisplayName(BoolValueDisplayName);
 	BoolValuePropertyRow->ToolTip(BoolValueToolTip);
 
-	IntValuePropertyRow->DisplayName(IntValueDisplayName);
-	IntValuePropertyRow->ToolTip(IntValueToolTip);
+	IntValuePropertyRow->SetDisplayName(IntValueDisplayName);
+	IntValuePropertyRow->SetToolTip(IntValueToolTip);
+	IntValuePropertyRow->Update();
 	FloatValuePropertyRow->ToolTip(FloatValueToolTip);
+
+	GUIDPropertyRow->DisplayName(GUIDDisplayName);
+	GUIDPropertyRow->ToolTip(GUIDToolTip);
 
 	// Refresh the view, without this some names/tooltips won't get refreshed
 	if (bForceRefresh && PropertyUtils.IsValid())