+++ Source/DlgSystemEditor/Private/DialogueEditor/DetailsPanel/Widgets/DialogueBase_CustomRowHelper.h
@@ -16,55 +16,55 @@ public:
 	FDialogueBase_CustomRowHelper(FDetailWidgetRow* InDetailWidgetRow, const TSharedPtr<IPropertyHandle>& InPropertyHandle);
 	virtual ~FDialogueBase_CustomRowHelper() {}
 
-	/** Sets the localized display name of the property. */
+	// Sets the localized display name of the property.
 	Self& SetDisplayName(const FText& InDisplayName)
 	{
 		DisplayName = InDisplayName;
 		return *this;
 	}
 
-	/** Sets the localized tooltip of the property. */
+	// Sets the localized tooltip of the property.
 	Self& SetToolTip(const FText& InToolTip)
 	{
 		ToolTip = InToolTip;
 		return *this;
 	}
 
-	/** Sets the visibility of this property. */
+	// Sets the visibility of this property.
 	Self& SetVisibility(const TAttribute<EVisibility>& InVisibility)
 	{
 		DetailWidgetRow->Visibility(InVisibility);
 		return *this;
 	}
 
-	/** Set the optional utils */
+	// Set the optional utils
 	Self& SetPropertyUtils(const TSharedPtr<IPropertyUtilities>& Utils)
 	{
 		PropertyUtils = Utils;
 		return *this;
 	}
 
-	/** Update the full property row. */
+	// Update the full property row.
 	void Update();
 
 protected:
-	/** Internal update method that must be implemented. Called after Update finishes. */
+	// Internal update method that must be implemented. Called after Update finishes.
 	virtual void UpdateInternal() = 0;
 
 protected:
-	/** The custom widget row this represents */
+	// The custom widget row this represents
 	FDetailWidgetRow* DetailWidgetRow = nullptr;
 
-	/** Helpers class  */
+	// Helpers class
 	TSharedPtr<IPropertyUtilities> PropertyUtils;
 
-	/** The Property handle of what this row represents */
+	// The Property handle of what this row represents
 	TSharedPtr<IPropertyHandle> PropertyHandle;
 
-	/** The NameContent Widget. */
+	// The NameContent Widget.
 	TSharedPtr<SWidget> NameContentWidget;
 
-	/** Texts used for this property row. */
+	// Texts used for this property row.
 	FText DisplayName;
 	FText ToolTip;
 };